examples/p2p-webrtc/video-transform/server/requirements.txt
python-dotenv
fastapi[all]
uvicorn
aiortc
opencv-python
pipecat-ai[google,silero,webrtc,deepgram,openai,gstreamer,noisereduce]
pipecat-ai-small-webrtc-prebuilt
pipecat-ai-flows
pygobject-stubs
